+++ src/app/(members)/mitglieder/endproben-woche/constants.ts
@@ -1,0 +1,5 @@
+export const FINAL_WEEK_VIEW_PERMISSION_KEY =
+  "mitglieder.endprobenwoche" as const;
+
+export const FINAL_WEEK_MANAGE_PERMISSION_KEY =
+  "mitglieder.endprobenwoche.manage" as const;
