+++ src/app/(members)/mitglieder/endproben-woche/page.tsx
@@ -1,0 +1,7 @@
+import { redirect } from "next/navigation";
+
+export const dynamic = "force-dynamic";
+
+export default function FinalWeekIndexRedirect() {
+  redirect("/mitglieder/endproben-woche/dienstplan");
+}
